Fixed the nullpointerexception when deleting a shared transition

+++ src/dk/aau/cs/gui/DeleteSharedPlaceOrTransition.java
@@ -219,8 +219,10 @@ public class DeleteSharedPlaceOrTransition implements ActionListener{
 		if(deleteFromTemplates){
 			for(Template template : tab.allTemplates()){ // TODO: Get rid of pipe references somehow
 				TimedTransitionComponent transition = (TimedTransitionComponent)template.guiModel().getTransitionByName(sharedTransition.name());
-				undoManager.addEdit(new DeleteTimedTransitionCommand(transition, transition.underlyingTransition().model(), template.guiModel(), tab.drawingSurface()));
-				transition.delete();
+				if(transition == null){
+					undoManager.addEdit(new DeleteTimedTransitionCommand(transition, transition.underlyingTransition().model(), template.guiModel(), tab.drawingSurface()));
+					transition.delete();
+				}
 			}
 			tab.drawingSurface().repaint();
 			sharedTransitionsListModel.removeElement(sharedTransition);
